feat: Response 구조 수정
