--- src/main/java/com/monglife/core/enums/error/ErrorCode.java
@@ -1,7 +1,0 @@
-package com.monglife.core.enums.error;
-
-public interface ErrorCode {
-    Integer getHttpStatus();
-    String getCode();
-    String getMessage();
-}
