+++ src/main/java/com/monglife/core/enums/response/Response.java
@@ -1,0 +1,12 @@
+package com.monglife.core.enums.response;
+
+import com.monglife.core.dto.response.ResponseDto;
+
+import java.util.Map;
+
+public interface Response {
+
+    ResponseDto<Map<String, Object>> toResponse();
+
+    <T> ResponseDto<T> toResponseDto(T result);
+}
