+++ src/main/java/com/monglife/core/enums/role/RoleCode.java
@@ -1,11 +1,13 @@
 package com.monglife.core.enums.role;
 
 public enum RoleCode {
+
     NORMAL("NORMAL", "일반회원"),
     ADMIN("ADMIN", "관리자"),
     ;
 
     private final String role;
+
     private final String name;
 
     RoleCode(String role, String name) {
